[COMMON] g2d: fix unsigned datatype
'get_unused_fd_flags' is stored in unsigned datatype,
so could not be handled on error.

+++ drivers/gpu/exynos/g2d/g2d_fence.c
@@ -176,7 +176,7 @@ struct sync_file *g2d_create_release_fence(struct g2d_device *g2d_dev,
 
 	for (i = 0; i < data->num_release_fences; i++) {
 		release_fences[i] = get_unused_fd_flags(O_CLOEXEC);
-		if (release_fences[i] < 0) {
+		if ((int)release_fences[i] < 0) {
 			ret = release_fences[i];
 			while (i-- > 0)
 				put_unused_fd(release_fences[i]);
